requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LangChain Core
langchain==0.1.20
langchain-community==0.0.38

# LLM and Embeddings
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4  # Compatible with both transformers and sentence-transformers
accelerate==0.25.0

# Vector Store
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU support

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
unstructured==0.11.6
pdfplumber==0.10.3

# Utilities
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Optional: For production deployment
gunicorn==21.2.0
